requirements.txt
langchain>=0.2.11
langgraph>=0.2.16
langchain-aws>=0.1.10
streamlit>=1.36.0
python-dotenv>=1.0.1
boto3>=1.34.100
pydantic>=2.7.0
typing-extensions>=4.12.0
